+++ changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.0.12] - 2023-01-15
+
+## Added
+
+- add support for `TRACEPARENT` environment variable; format matches the [w3 format](https://www.w3.org/TR/trace-context-1/).
+
+## Fixed
+
+- flag environment variable overriding wasn't handling boolean values correctly; any value was considered `true`.
+
 ## [0.0.11] - 2023-01-09
 
 ## Fixed